# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Film Master API",
    "description": "My first MERN project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer":{
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/api/film": {
      "get": {
        "summary": "Get all films",
        "security": [
          "Bearer"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/film/{id}": {
      "get": {
        "summary": "Get film by id",
        "security": [
          "Bearer"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/film"
            }
          }
        ],
        "security": [
          "Bearer"
        ],
        "summary": "Partial update",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/api/film/{id}/edit": {
      "put": {
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/film"
            }
          }
        ],
        "security": [
          "Bearer"
        ],
        "summary": "Edit film",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/api/film/{id}/delete": {
      "delete": {
        "summary": "Delete Film by id",
        "security": [
          "Bearer"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "summary": "Login",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }

    },
    "/api/film/add": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/film"
            }
          }
        ],
        "security": [
          "Bearer"
        ],
        "summary": "Add Film",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    }
  },
  "definitions": {
    "user": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "film": {
      "properties": {
        "name": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        },
        "isWatched": {
          "type": "boolean"
        }
      }
    }
  }
}